+++ compiler-plugin/src/main/kotlin/tech/mappie/generation/ConstructorCallMappingConstructor.kt
@@ -2,26 +2,33 @@ package tech.mappie.generation
 
 import org.jetbrains.kotlin.ir.IrFileEntry
 import org.jetbrains.kotlin.ir.builders.*
+import org.jetbrains.kotlin.ir.declarations.IrClass
 import org.jetbrains.kotlin.ir.declarations.IrFunction
 import org.jetbrains.kotlin.ir.declarations.IrValueParameter
 import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.expressions.IrFunctionAccessExpression
+import org.jetbrains.kotlin.ir.types.isNullable
 import org.jetbrains.kotlin.ir.util.constructors
+import org.jetbrains.kotlin.ir.util.functions
 import org.jetbrains.kotlin.ir.util.isObject
 import tech.mappie.MappieIrRegistrar
 import tech.mappie.MappieIrRegistrar.Companion.context
 import tech.mappie.mappieTerminate
 import tech.mappie.resolving.ConstructorCallMapping
+import tech.mappie.resolving.classes.MappieViaClass
+import tech.mappie.resolving.classes.MappieViaGeneratedEnumClass
 import tech.mappie.resolving.classes.*
 import tech.mappie.resolving.classes.targets.MappieFunctionTarget
 import tech.mappie.resolving.classes.targets.MappieSetterTarget
+import tech.mappie.resolving.classes.targets.MappieTarget
 import tech.mappie.resolving.classes.targets.MappieValueParameterTarget
-import tech.mappie.util.location
-import tech.mappie.util.referenceFunctionLet
-import tech.mappie.util.targetType
+import tech.mappie.util.*
 
-class ConstructorCallMappingConstructor(private val mapping: ConstructorCallMapping, declaration: IrFunction)
-    : MappingConstructor( declaration) {
+class ConstructorCallMappingConstructor(
+    private val generated: List<IrClass>,
+    private val mapping: ConstructorCallMapping,
+    declaration: IrFunction,
+) : MappingConstructor(declaration) {
 
     override fun construct(scope: Scope) =
         context.blockBody(scope) {
@@ -29,7 +36,7 @@ class ConstructorCallMappingConstructor(private val mapping: ConstructorCallMapp
                 mapping.mappings.toList().forEach { (target, source) ->
                     if (target is MappieValueParameterTarget && source.single() !is DefaultArgumentSource) {
                         val index = mapping.symbol.owner.valueParameters.indexOf(target.value)
-                        putValueArgument(index, generateValueArgument(file, source.single(), declaration))
+                        putValueArgument(index, generateValueArgument(file, source.single(), target, declaration))
                     }
                 }
             }
@@ -41,13 +48,13 @@ class ConstructorCallMappingConstructor(private val mapping: ConstructorCallMapp
                     is MappieSetterTarget -> {
                         +irCall(target.value.setter!!).apply {
                             dispatchReceiver = irGet(variable)
-                            putValueArgument(0, generateValueArgument(file, source.single(), declaration))
+                            putValueArgument(0, generateValueArgument(file, source.single(), target, declaration))
                         }
                     }
                     is MappieFunctionTarget -> {
                         +irCall(target.value).apply {
                             dispatchReceiver = irGet(variable)
-                            putValueArgument(0, generateValueArgument(file, source.single(), declaration))
+                            putValueArgument(0, generateValueArgument(file, source.single(),target, declaration))
                         }
                     }
                     is MappieValueParameterTarget -> { /* Applied as a constructor call argument */ }
@@ -57,9 +64,14 @@ class ConstructorCallMappingConstructor(private val mapping: ConstructorCallMapp
             +irReturn(irGet(variable))
         }
 
-    private fun IrBuilderWithScope.generateValueArgument(file: IrFileEntry, source: ObjectMappingSource, function: IrFunction): IrExpression {
+    private fun IrBuilderWithScope.generateValueArgument(
+        file: IrFileEntry,
+        source: ObjectMappingSource,
+        target: MappieTarget,
+        function: IrFunction
+    ): IrExpression {
         return when (source) {
-            is ResolvedSource -> generateResolvedValueArgument(source, function)
+            is ResolvedSource -> generateResolvedValueArgument(source, target, function)
             is PropertySource -> generatePropertyValueArgument(file, source, function.valueParameters)
             is ExpressionSource -> generateExpressionValueArgument(source, function.valueParameters)
             is ValueSource -> source.value
@@ -67,26 +79,65 @@ class ConstructorCallMappingConstructor(private val mapping: ConstructorCallMapp
         }
     }
 
-    private fun IrBuilderWithScope.generateResolvedValueArgument(source: ResolvedSource, function: IrFunction): IrFunctionAccessExpression {
+    private fun IrBuilderWithScope.generateResolvedValueArgument(
+        source: ResolvedSource,
+        target: MappieTarget,
+        function: IrFunction
+    ): IrFunctionAccessExpression {
         val getter = irCall(source.property.function).apply {
             dispatchReceiver = irGet(source.property.holder)
         }
-        return source.via?.let { (clazz, via) ->
-            irCall(via).apply {
-                dispatchReceiver = when {
-                    clazz.isObject -> irGetObject(clazz.symbol)
-                    else -> {
-                        val constructor = clazz.constructors.firstOrNull { it.valueParameters.isEmpty() }
-                        if (constructor != null) {
-                            irCallConstructor(constructor.symbol, emptyList())
-                        } else {
-                            mappieTerminate("Resolved mapping via ${clazz.name.asString()}, but it does not have a constructor without arguments", location(function))
-                        }
-                    }
+        return source.via?.let { via ->
+            when (via) {
+                is MappieViaClass -> {
+                    generateVia(source, target, via.clazz, function, getter)
                 }
-                putValueArgument(0, getter)
+                is MappieViaGeneratedEnumClass -> {
+                    val clazz = generated.find { it.name == via.name }
+                    generateVia(source, target, clazz!!, function, getter)
+                }
             }
         } ?: getter
+    }
+
+    private fun IrBuilderWithScope.generateVia(
+        source: ResolvedSource,
+        target: MappieTarget,
+        via: IrClass,
+        function: IrFunction,
+        getter: IrFunctionAccessExpression,
+    ): IrFunctionAccessExpression {
+        val viaFunction = when {
+            source.type.isList() && target.type.isList() ->
+                via.functions.first { it.isMappieMapListFunction() }
+
+            source.type.isSet() && target.type.isSet() ->
+                via.functions.first { it.isMappieMapSetFunction() }
+
+            source.type.isNullable() && target.type.isNullable() ->
+                via.functions.first { it.isMappieMapNullableFunction() }
+
+            else ->
+                via.functions.first { it.isMappieMapFunction() }
+        }
+
+        return irCall(viaFunction).apply {
+            dispatchReceiver = when {
+                via.isObject -> irGetObject(via.symbol)
+                else -> {
+                    val constructor = via.constructors.firstOrNull { it.valueParameters.isEmpty() }
+                    if (constructor != null) {
+                        irCallConstructor(constructor.symbol, emptyList())
+                    } else {
+                        mappieTerminate(
+                            "Resolved mapping via ${via.name.asString()}, but it does not have a constructor without arguments",
+                            location(function)
+                        )
+                    }
+                }
+            }
+            putValueArgument(0, getter)
+        }
     }
 
     private fun IrBuilderWithScope.generatePropertyValueArgument(file: IrFileEntry, source: PropertySource, parameters: List<IrValueParameter>): IrExpression {