+++ compiler-plugin/src/main/kotlin/tech/mappie/generation/MappieGeneration.kt
@@ -1,0 +1,10 @@
+package tech.mappie.generation
+
+import org.jetbrains.kotlin.ir.declarations.IrFunction
+import tech.mappie.resolving.Mapping
+import tech.mappie.resolving.classes.MappieVia
+
+class MappieGeneration(
+    val mappings: Map<IrFunction, Mapping>,
+    val generated: Set<MappieVia>,
+)