+++ compiler-plugin/src/main/kotlin/tech/mappie/validation/problems/UnsafePlatformTypeAssignmentProblems.kt
@@ -17,9 +17,9 @@ class UnsafePlatformTypeAssignmentProblems(
     private val mappings: List<Pair<MappieTarget, ObjectMappingSource>>,
 ) {
 
-    fun all(): List<Problem> = mappings.map { warning(it.first, it.second) }.filterNotNull()
+    fun all(): List<Problem> = mappings.map { validate(it.first, it.second) }.filterNotNull()
 
-    private fun warning(target: MappieTarget, source: ObjectMappingSource): Problem? {
+    private fun validate(target: MappieTarget, source: ObjectMappingSource): Problem? {
         val sourceTypeString = source.type.removeAnnotations().dumpKotlinLike()
         val targetTypeString = targetType.dumpKotlinLike()
         val targetString = "$targetTypeString::${target.name.asString()}"