+++ compiler-plugin/src/test/kotlin/tech/mappie/testing/enums/NestedNestedGeneratedEnumTest.kt
@@ -1,4 +1,4 @@
-package tech.mappie.testing
+package tech.mappie.testing.enums
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -6,12 +6,22 @@ import org.junit.jupiter.api.io.TempDir
 import tech.mappie.testing.compilation.KotlinCompilation
 import tech.mappie.testing.compilation.KotlinCompilation.ExitCode
 import tech.mappie.testing.compilation.SourceFile.Companion.kotlin
+import tech.mappie.testing.loadObjectMappieClass
 import java.io.File
 
-class MapperClassCanMultipleMapFunctionsTest {
+class NestedNestedGeneratedEnumTest {
 
-    data class Input(val text: String)
-    data class Output(val text: String, val int: Int)
+    data class Input(val nested: NestedInput)
+
+    data class NestedInput(val nested: NestedNestedEnumInput)
+
+    enum class NestedNestedEnumInput { A, B, C }
+
+    data class Output(val nested: NestedOutput)
+
+    data class NestedOutput(val nested: NestedNestedEnumOutput)
+
+    enum class NestedNestedEnumOutput { A, B, C }
 
     @TempDir
     lateinit var directory: File
@@ -24,16 +34,10 @@ class MapperClassCanMultipleMapFunctionsTest {
                     kotlin("Test.kt",
                         """
                         import tech.mappie.api.ObjectMappie
-                        import tech.mappie.testing.MapperClassCanMultipleMapFunctionsTest.*
+                        import tech.mappie.testing.enums.NestedNestedGeneratedEnumTest.*
     
-                        class Mapper(private val int: Int) : ObjectMappie<Input, Output>() {
-                            fun map(value: String) = value
-
-                            override fun map(from: Input) = mapping {
-                                Output::int fromValue int
-                            }
-
-                            fun map(value: Int) = value
+                        class Mapper : ObjectMappie<Input, Output>() {
+                            object NestedMapper : ObjectMappie<NestedInput, NestedOutput>()
                         }
                         """
                     )
@@ -47,11 +51,10 @@ class MapperClassCanMultipleMapFunctionsTest {
                 .loadObjectMappieClass<Input, Output>("Mapper")
                 .constructors
                 .first()
-                .call(10)
+                .call()
 
-            assertThat(mapper.map(Input("test")))
-                .isEqualTo(Output("test", 10))
+            assertThat(mapper.map(Input(NestedInput(NestedNestedEnumInput.A))))
+                .isEqualTo(Output(NestedOutput(NestedNestedEnumOutput.A)))
         }
     }
-
 }