+++ compiler-plugin/src/test/kotlin/tech/mappie/testing/enums/ObjectWithNestedEnumTest.kt
@@ -19,6 +19,37 @@ class ObjectWithNestedEnumTest {
     lateinit var directory: File
 
     @Test
+    fun `map data classes with nested enum using object InnerMapper without declaring nested mapping should succeed`() {
+        KotlinCompilation(directory).apply {
+            sources = buildList {
+                add(
+                    kotlin("Test.kt",
+                        """
+                        import tech.mappie.api.ObjectMappie
+                        import tech.mappie.api.EnumMappie
+                        import tech.mappie.testing.enums.ObjectWithNestedEnumTest.*
+    
+                        class Mapper : ObjectMappie<Input, Output>()
+                        """
+                    )
+                )
+            }
+        }.compile {
+            assertThat(exitCode).isEqualTo(ExitCode.OK)
+            assertThat(messages).isEmpty()
+
+            val mapper = classLoader
+                .loadObjectMappieClass<Input, Output>("Mapper")
+                .constructors
+                .first()
+                .call()
+
+            assertThat(mapper.map(Input(InnerEnum.A)))
+                .isEqualTo(Output(OuterEnum.A))
+        }
+    }
+
+    @Test
     fun `map data classes with nested enum using object InnerMapper without declaring mapping should succeed`() {
         KotlinCompilation(directory).apply {
             sources = buildList {